app/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pymysql
passlib
slowapi